# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Recipes api
  description: API for recipes project
  version: 1.0.1
servers:
# local:
- url: "http://127.0.0.1:3000"
# remote:
# - url: "http://132.72.65.211"
  description: SwaggerHub API Auto Mocking hey
...
paths:
...
  /recipes/recipeExtendedInfo/{recipe_id}/{userId}:
    get:
      tags:
      - Recipes
      summary: get full recipe by id
      parameters:
      - name: recipe_id
        in: path
        required: true
        schema:
          type: integer

      - name: recipe_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: 'recipe'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullRecipe'
  /recipes/recipePreview:
    post:
      tags:
      - Recipes
      summary: show preview of recipe
      description: show preview of recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeId'
      responses: 
        '200':
          description: '#/components/schemas/Recipe'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
...
components:
  schemas:
...
    Recipe:
      type: object
      properties:
        RecipeImg:
          type: string
          description: the name of the recipe
        RecipeName:
          type: string
          description: the image of the recipe
          format: url
        CookingTime:
          type: integer
          description: the cooking time
        GlutenFree:
          type: boolean
          description: is gluten free or not
        isVegan:
          type: boolean
          description: is vegan or not
        isVegetarian:
          type: boolean
          description: is vegeterian or not
        isCicked:
          type: boolean
          description: is clicked by the user or not
        isFavorite:
          type: boolean
          description: is Favorite by the user or not
...
    FullRecipe:
      required:
      - RecipeImg
      - RecipeName
      - CookingTime
      - GlutenFree
      - isVegan
      - isVegetarian
      - isClicked
      - isFavorite
      - ingredients
      - instructions
      - servings
      type: object
      properties:
        RecipeImg:
          type: string
          description: the name of the recipe
        RecipeName:
          type: string
          description: the image of the recipe
          format: url
        CookingTime:
          type: integer
          description: the cooking time
        GlutenFree:
          type: boolean
          description: is gluten free or not
        isVegan:
          type: boolean
          description: is vegan or not
        isVegetarian:
          type: boolean
          description: is vegeterian or not
        isCicked:
          type: boolean
          description: is clicked by the user or not
        isFavorite:
          type: boolean
          description: is Favorite by the user or not
        ingredients:
          type: string
          description: the ingredients needed for the recipe
        instructions:
          type: string
          description: the instructions for the recipe
        servings:
          type: string
          description: the total servings of the recipe
    RecipeId:
        type: string
...